restructure functions to avoid crossed imports
